web/requirements.txt
flask
flask-restful
spacy
urllib3
pyyaml
tabulate
bs4
openpyxl
requests
xlsxwriter
colorama
hjson
termcolor
idna
pandas
Pillow
numpy
scipy
scikit-learn
scikit-image
scikit-plot
pickleshare
cloudpickle
packaging